Revert "refactoring QVE saved generation component (#102)"
This reverts commit 8f8cc8334e46f80c0fffc64b330c64dbcbd24f0b.
