--- add-in/src/pages/qve/iconFunc.tsx
@@ -1,13 +1,0 @@
-import { AiOutlineAlignLeft, AiOutlineQuestion, AiOutlineHighlight } from 'react-icons/ai';
-
-export const iconFunc = (generation_type: string) => {
-    switch (generation_type) {
-        case 'Completion':
-            return<AiOutlineAlignLeft />;
-        case 'Question':
-            return <AiOutlineQuestion />;
-        case 'Keywords':
-            return <AiOutlineHighlight />;
-    }
-};
-
