--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "eslint": "^9.11.1"
-  }
-}
